Rename prefer-textcontent to prefer-text-content
