+++ docs/rules/prefer-text-content.md
