+++ readme.md
@@ -73,7 +73,7 @@ Configure it in `package.json`.
 			"unicorn/prefer-node-append": "error",
 			"unicorn/prefer-query-selector": "error",
 			"unicorn/prefer-node-remove": "error",
-			"unicorn/prefer-textcontent": "error"
+			"unicorn/prefer-text-content": "error"
 		}
 	}
 }
@@ -111,7 +111,7 @@ Configure it in `package.json`.
 - [prefer-node-append](docs/rules/prefer-node-append.md) - Prefer `append` over `appendChild`. *(fixable)*
 - [prefer-query-selector](docs/rules/prefer-query-selector.md) - Prefer `querySelector` over `getElementById`, `querySelectorAll` over `getElementsByClassName` and `getElementsByTagName`. *(partly fixable)*
 - [prefer-node-remove](docs/rules/prefer-node-remove.md) - Prefer `remove` over `parentNode.removeChild` and `parentElement.removeChild`. *(fixable)*
-- [prefer-textcontent](docs/rules/prefer-textcontent.md) - Prefer `textContent` over `innerText`. *(fixable)*
+- [prefer-text-content](docs/rules/prefer-text-content.md) - Prefer `textContent` over `innerText`. *(fixable)*
 
 
 ## Recommended config
