+++ rules/prefer-text-content.js
